feat(core): add `rag`
